+++ CHANGELOG.md
@@ -1,3 +1,44 @@
+### 3.1.49
+#### Games added
+  - Gladio Mori
+  - Slipstream: Rogue Space
+  - Panicore
+  - Magicraft
+  - Another Crab's Treasure
+  - Bopl Battle
+  - Vertigo 2
+  - Against the Storm
+  - Lycans
+  - Castle Story
+  - Balatro
+
+#### General fixes
+- Better online mod filtering
+- Filter local mod list by last updated date
+  - This will only apply to mods updated as of this release
+- Smaller fixes / tweaks
+
+#### BONEWORKS / BONELAB
+- BONEWORKS and BONELAB now have "recommended" mod loader versions to aid the experience
+
+#### Titanfall 2 / Northstar
+- The Northstar mod loader package installs correctly again
+
+### 3.1.48
+- Games added:
+  - Content Warning
+  - Plasma
+  - MS store support for Bomb Rush Cyberfunk
+- Profile code exports now have a size restriction
+- Fixes for Voices of the Void
+- Other general fixes/improvements
+
+_Contributions:_
+- _anttimaki_
+- _cspotcode_
+- _Oksamies_
+- _ethangreen-dev_
+
 ### 3.1.47
 - Significant performance improvements
 - Games added: